- serial part multi version support changes
