+++ modules/sde-core/pom.xml
@@ -2,8 +2,8 @@
 <!-- 
 /******************************************************************************** 
 * Copyright (c) 2022 BMW GmbH 
-* Copyright (c) 2022,2023 T-Systems International GmbH 
-* Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation 
+* Copyright (c) 2022, 2024 T-Systems International GmbH 
+* Copyright (c) 2022, 2024 Contributors to the Eclipse Foundation 
 * 
 * See the NOTICE file(s) distributed with this work for additional 
 * information regarding copyright ownership. 
@@ -128,7 +128,7 @@
 
 		<dependency>
 			<groupId>org.eclipse.tractusx</groupId>
-			<artifactId>serial-part-typization</artifactId>
+			<artifactId>serial-part</artifactId>
 			<version>0.0.1</version>
 		</dependency>
