--- modules/sde-submodules/serial-part-typization/src/main/java/org/eclipse/tractusx/sde/submodels/spt/SerialPartTypization101.java
@@ -1,58 +1,0 @@
-/********************************************************************************
- * Copyright (c) 2022 Critical TechWorks GmbH
- * Copyright (c) 2022 BMW GmbH
- * Copyright (c) 2022, 2023 T-Systems International GmbH
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ********************************************************************************/
-package org.eclipse.tractusx.sde.submodels.spt;
-
-import jakarta.annotation.PostConstruct;
-import org.eclipse.tractusx.sde.common.extensions.SubmodelExtension;
-import org.eclipse.tractusx.sde.common.model.Submodel;
-import org.springframework.stereotype.Component;
-
-import java.io.InputStream;
-
-@Component
-public class SerialPartTypization101 extends SubmodelExtension {
-
-    private Submodel submodel = null;
-
-    @PostConstruct
-    public void init() {
-
-        String resource = "serial-part-typization_1.0.1.json";
-        // this is the path within the jar file
-        InputStream input = this.getClass().getResourceAsStream("/resources/" + resource);
-        if (input == null) {
-            // this is how we load file within editor (eg eclipse)
-            input = this.getClass().getClassLoader().getResourceAsStream(resource);
-        }
-
-        submodel = loadSubmodel(input);
-
-        submodel.addProperties("tableName", "aspect101");
-    }
-
-    @Override
-    public Submodel submodel() {
-        return submodel;
-    }
-
-
-}
