+++ modules/sde-submodules/serial-part/pom.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 /********************************************************************************
- * Copyright (c) 2022 BMW GmbH
- * Copyright (c) 2022, 2023 T-Systems International GmbH
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024 T-Systems International GmbH
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -31,9 +30,9 @@
 		<version>0.0.1</version>
 		<relativePath>../../../</relativePath>
 	</parent>
-	<artifactId>serial-part-typization</artifactId>
-	<name>serial-part-typization</name>
-	<description>serial-part-typization</description>
+	<artifactId>serial-part</artifactId>
+	<name>serial-part</name>
+	<description>serial-part</description>
 
 	<dependencies>
 		<dependency>
