+++ modules/sde-submodules/serial-part/src/main/resources/serial-part-v1.0.0.json
