+++ modules/sde-submodules/serial-part/src/main/resources/serial-part-v3.0.0.json
@@ -2,10 +2,10 @@
 	"$schema": "https://json-schema.org/draft/2019-09/schema",
 	"$id": "http://example.com/example.json",
 	"type": "array",
-	"id": "serialpart_101",
+	"id": "serialpart_300",
 	"idShort": "serialPart",
-	"version": "1.0.1",
-	"semantic_id": "urn:bamm:io.catenax.serial_part:1.0.1#SerialPart",
+	"version": "3.0.0",
+	"semantic_id": "urn:samm:io.catenax.serial_part:3.0.0#SerialPart",
 	"title": "Serial Part",
 	"shortDescription": "BoM As-Built - Submodel SerialPart",
 	"description": "A serialized part is an instantiation of a (design-) part, where the particular instantiation can be uniquely identified by means of a serial numbers or a similar identifier (e.g. VAN) or a combination of multiple identifiers (e.g. combination of manufacturer, date and number)",
@@ -16,7 +16,10 @@
 			"manufacturing_date",
 			"manufacturer_part_id",
 			"name_at_manufacturer",
-			"classification"
+			"catenax_site_id",
+			"function",
+			"classification_standard",
+			"classification_ID"
 		],
 		"dependentRequired": {
 			"optional_identifier_key": [
@@ -56,6 +59,7 @@
 				],
 				"format": "date-time",
 				"title": "Manufacturing Date",
+				"pattern": "^(?:[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?Z|[0-9]{4}-[0-9]{2}-[0-9]{2}(?:T[0-9]{2}:[0-9]{2}:[0-9]{2}(?:[.][0-9]+)?(?:Z|[+-][0-9]{2}:[0-9]{2}))?)$",
 				"description": "Timestamp of the manufacturing date as the final step in production process (e.g. final quality check, ready-for-shipment event)",
 				"examples": [
 					"2022-02-04T14:48:54.709Z"
@@ -264,6 +268,34 @@
 					"DEU"
 				]
 			},
+			"catenax_site_id": {
+				"type": [
+					"string"
+				],
+				"title": "CatenaX site Id",
+				"minLength": 1,
+				"pattern": "^BPNS[a-zA-Z0-9]{12}$",
+				"description": "The identifier of the site according to Catena-X BPDM. The catenaXsiteId must be a valid Catena-X BPN. The BPN is a unique, unchangeable identifier for Business Partners / company locations from foundation to closure, regardless of the different business relationships / structures between or within the Business Partners or company locations.",
+				"examples": [
+					"BPNS1234567890ZZ"
+				]
+			},
+			"function": {
+				"type": [
+					"string"
+				],
+				"enum": [
+					"production",
+					"warehouse",
+					"spare part warehouse"
+				],
+				"title": "function",
+				"minLength": 1,
+				"description": "The function of the site in relation to the part(i.e. the activity within the value chain of the part that is performed at the site)",
+				"examples": [
+					"production"
+				]
+			},
 			"manufacturer_part_id": {
 				"type": [
 					"string"
@@ -284,24 +316,6 @@
 				"description": "Part ID as assigned by the manufacturer of the part. The Part ID identifies the part (as designed) in the customer`s dataspace. The Part ID does not reference a specific instance of a part and thus should not be confused with the serial number.",
 				"examples": [
 					"37754B7-76"
-				]
-			},
-			"classification": {
-				"type": [
-					"string"
-				],
-				"enum": [
-					"product",
-					"raw material",
-					"software",
-					"assembly",
-					"tool",
-					"component"
-				],
-				"title": "Classification",
-				"description": "The classification of the part type according to STEP standard definition.",
-				"examples": [
-					"component"
 				]
 			},
 			"name_at_manufacturer": {
@@ -352,6 +366,39 @@
 				"examples": [
 					"Value"
 				]
+			},
+			"classification_standard": {
+				"type": [
+					"string"
+				],
+				"minLength": 1,
+				"title": "Classification Standard",
+				"description": "Identified classification standards that align to the Catena-X needs.",
+				"examples": [
+					"IEC"
+				]
+			},
+			"classification_ID": {
+				"type": [
+					"string"
+				],
+				"minLength": 1,
+				"title": "Classification ID",
+				"description": "The classification ID of the part type according to the corresponding standard definition mentioned in the key value pair.",
+				"examples": [
+					"61360- 2:2012"
+				]
+			},
+			"classification_description": {
+				"type": [
+					"string",
+					"null"
+				],
+				"title": "Classification Description",
+				"description": "Optional property describing the classification standard.",
+				"examples": [
+					"Standard data element types with associated classification scheme for electric components."
+				]
 			}
 		}
 	},
@@ -361,32 +408,29 @@
 			"part_instance_id": "NO-159040131155901488695376",
 			"manufacturing_date": "2022-02-04T14:48:54.709Z",
 			"manufacturing_country": "DEU",
+			"catenax_site_id": "BPNS1234567890ZZ",
+			"function": "production",
 			"manufacturer_part_id": "37754B7-76",
 			"customer_part_id": "37754B7-76",
-			"classification": "component",
 			"name_at_manufacturer": "Sensor",
 			"name_at_customer": "Sensor",
 			"optional_identifier_key": "van",
-			"optional_identifier_value": "Value"
+			"optional_identifier_value": "Value",
+			"classification_standard": "IEC",
+			"classification_ID": "61360- 2:2012 ",
+			"classification_description": "Standard data element types with associated classification scheme for electric components."
 		}
 	],
 	"addOn": {
 		"identifier": "uuid",
 		"specificAssetIdsSpecs": {
 			"partInstanceId": "part_instance_id",
-			"manufacturerPartId": "manufacturer_part_id",
-			"optionalIdentifier": [
-				{
-					"key": "optional_identifier_key",
-					"value" :"optional_identifier_value"
-				}
-			]
+			"manufacturerPartId": "manufacturer_part_id"
 		},
 		"shortIdSpecs": [
 			"name_at_manufacturer",
 			"part_instance_id",
 			"manufacturer_part_id"
-			
 		],
 		"bpnDiscoverySpecs": {
 			"manufacturerPartId": "manufacturer_part_id"
@@ -394,11 +438,11 @@
 		"responseTemplate": {
 			"localIdentifiers": [
 				{
-					"key": "PartInstanceId",
+					"key": "partInstanceId",
 					"value": "${part_instance_id}"
 				},
 				{
-					"key": "ManufacturerPartID",
+					"key": "ManufacturerPartId",
 					"value": "${manufacturer_part_id}"
 				},
 				{
@@ -406,19 +450,31 @@
 					"value": "${optional_identifier_value}"
 				},
 				{
-					"key": "ManufacturerID",
+					"key": "ManufacturerId",
 					"value": "${manufacturerId}"
 				}
 			],
 			"manufacturingInformation": {
 				"date": "${manufacturing_date}",
-				"country": "${manufacturing_country}"
+				"country": "${manufacturing_country}",
+				"sites": [
+					{
+						"catenaXsiteId": "${catenax_site_id}",
+						"function": "${function}"
+					}
+				]
 			},
 			"catenaXId": "${uuid}",
 			"partTypeInformation": {
+				"partClassification": [
+					{
+						"classificationStandard": "${classification_standard}",
+						"classificationID": "${classification_ID}",
+						"classificationDescription": "${classification_description}"
+					}
+				],
 				"manufacturerPartId": "${manufacturer_part_id}",
 				"customerPartId": "${customer_part_id}",
-				"classification": "${classification}",
 				"nameAtManufacturer": "${name_at_manufacturer}",
 				"nameAtCustomer": "${name_at_customer}"
 			}
