+++ pom.xml
@@ -2,8 +2,8 @@
 <!--
 /********************************************************************************
  * Copyright (c) 2022 BMW GmbH
- * Copyright (c) 2022, 2023 T-Systems International GmbH
- * Copyright (c) 2022, 2023 Contributors to the Eclipse Foundation
+ * Copyright (c) 2022, 2024 T-Systems International GmbH
+ * Copyright (c) 2022, 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -63,7 +63,7 @@
 		<module>modules/sde-core</module>
 		<module>modules/pcf-exchange</module>
 		
-		<module>modules/sde-submodules/serial-part-typization</module>
+		<module>modules/sde-submodules/serial-part</module>
 		<module>modules/sde-submodules/batch</module>
 		<module>modules/sde-submodules/assembly-part-relationship</module>
 		<module>modules/sde-submodules/part-as-planned</module>
@@ -71,7 +71,6 @@
 		<module>modules/sde-submodules/part-site-information-as-planned</module>
 		<module>modules/sde-submodules/single-level-usage-as-built</module>
 		<module>modules/sde-submodules/pcf</module>
-		
 	</modules>
 
 	<dependencies>
